cloudfunction/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.16.0
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.15.0
openai>=1.42.0
google-cloud-aiplatform==1.67.0


